services/booking_service/requirements.txt
kafka-python
psycopg2-binary
sqlalchemy
pydantic